+++ iqm_qaoa/CHANGELOG.rst
@@ -2,6 +2,24 @@
 Changelog
 =========
 
+Version 1.10.0 (2025-07-21)
+===========================
+
+Bug fixes
+---------
+
+- Add a fallback routine to `_get_embedding` which allows it to embed larger problems on QPUs on which it would previously fail, by using a greedy algorithm for finding a Hamiltonian path in a graph.
+- Add a small test to check that it works.
+
+Version 1.9.0 (2025-07-21)
+==========================
+
+Features
+--------
+
+- Add a Jupyter notebook showing how one can control / influence which qubits on the QPU get selected for execution of the circuit.
+- Add input ``**kwargs`` to ``transpiled_circuit`` which get passed to the inner ``transpile`` call (from Qiskit).
+
 Version 1.8.0 (2025-07-09)
 ==========================